requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.40
PyYAML>=6.0
